[test] 전투 시스템 작업사항
플레이어 주위에 마우스 방향을 따라다니는 커서
무기변경시 커서에 표시 (임시)

+++ Assets/WIP/BattleTest_PEY/BattleTeamB.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public class BattleTeamB : MonoBehaviour, IDamageable
 {
-    public int hp = 100;
+    public float hp = 100;
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         hp -= damage;
         Debug.Log(hp);
